yolov11/requirements.txt
albucore==0.0.24
albumentations==2.0.8
alembic==1.18.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
blinker==1.9.0
cachetools==6.2.4
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cloudpickle==3.1.2
contourpy==1.3.3
cryptography==46.0.3
cycler==0.12.1
databricks-sdk==0.78.0
docker==7.1.0
fastapi==0.128.0
filelock==3.20.3
Flask==3.1.2
flask-cors==6.0.2
fonttools==4.61.1
fsspec==2026.1.0
gitdb==4.0.12
GitPython==3.1.46
google-auth==2.47.0
graphene==3.4.3
graphql-core==3.2.7
graphql-relay==3.2.0
gunicorn==23.0.0
h11==0.16.0
huey==2.6.0
idna==3.11
importlib_metadata==8.7.1
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
kiwisolver==1.4.9
Mako==1.3.10
MarkupSafe==3.0.3
matplotlib==3.10.8
mlflow==3.8.1
mlflow-skinny==3.8.1
mlflow-tracing==3.8.1
mpmath==1.3.0
networkx==3.6.1
numpy==2.4.1
opencv-contrib-python==4.13.0.90
opencv-python==4.13.0.90
opencv-python-headless==4.13.0.90
opentelemetry-api==1.39.1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
packaging==25.0
pandas==2.3.3
pillow==12.1.0
polars==1.37.1
polars-runtime-32==1.37.1
protobuf==6.33.4
psutil==7.2.1
pyarrow==22.0.0
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pyparsing==3.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
rsa==4.9.1
scikit-learn==1.8.0
scipy==1.17.0
setuptools==80.9.0
simsimd==6.5.12
six==1.17.0
smmap==5.0.2
SQLAlchemy==2.0.45
sqlparse==0.5.5
starlette==0.50.0
stringzilla==4.6.0
sympy==1.14.0
threadpoolctl==3.6.0
torch==2.9.1
torchvision==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
ultralytics==8.4.6
ultralytics-thop==2.0.18
urllib3==2.6.3
uvicorn==0.40.0
Werkzeug==3.1.5
zipp==3.23.0
